SR-15: Move generate outside test/ folder.
